// generated/src/app/api-models/pcp-auto-assign-hdr.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class PcpAutoAssignHdr {
  seqPcpAutoAssgn: number;
  lineOfBusiness: string;
  eventType: number;
  reqPcpThreshold: string;
  membReinstThreshold: string;
  diamondIdThreshold: string;
  familyAfflThreshold: string;
  reqPcpFlag: string;
  reqPcpOrder: number;
  membReinstFlag: string;
  membReinstOrder: number;
  membReinstDays: number;
  diamondIdFlag: string;
  diamondIdOrder: number;
  diamondIdDays: number;
  familyAfflFlag: string;
  familyAfflOrder: number;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}